app/requirements.txt
fastapi
uvicorn
requests
azure-cosmos
azure-identity